+++ src/main.cpp
@@ -2,7 +2,6 @@
 #include "graph.h"
 #include "astar.h"
 #include <unordered_map>
-//#include "cbs.h"
 #include <vector>
 
 class SimpleGraph : public Graph<int> {
@@ -10,7 +9,7 @@ public:
     std::vector<std::unordered_map<int, int>> weights;
     std::vector<std::vector<int>> g;
 
-    int get_cost(int a, int b) override {
+    double get_cost(int a, int b) override {
         return weights[a][b];
     }
